joiShemas moved to middlewares/, loggers in one file, Article.remove instead of Arcticle.findByIdAndDelete
